+++ src/main/java/com/almostreliable/kubeio/schema/VatRecipeSchema.java
@@ -1,7 +1,6 @@
 package com.almostreliable.kubeio.schema;
 
-import com.enderio.machines.common.blocks.vat.FermentingRecipe;
-import com.enderio.machines.data.recipes.FermentingRecipeProvider;
+import com.enderio.enderio.content.machines.vat.FermentingRecipe;
 import dev.latvian.mods.kubejs.recipe.RecipeKey;
 import dev.latvian.mods.kubejs.recipe.component.*;
 import dev.latvian.mods.kubejs.recipe.schema.RecipeSchema;
@@ -11,7 +10,7 @@ import net.neoforged.neoforge.fluids.FluidStack;
 import net.neoforged.neoforge.fluids.crafting.SizedFluidIngredient;
 
 /**
- * See {@link FermentingRecipe.Serializer} and {@link FermentingRecipeProvider}.
+ * See {@link FermentingRecipe.Serializer}.
  */
 public interface VatRecipeSchema {